Registration form
Simon, make sure to add the path where your node.js program is collecting data and inserting it in the database.
So far, there is no way to know on if the data has been inserted. Let's try to get to insert data first, then we'll worry about sending data back to let the client know if the submission has been completed.